# Fix perf gap in thread safe prediction