As in https://github.com/dmlc/xgboost/pull/6648 local `RegTree::FVec` storage was introduced to preserve thread safety each training iteration requires buffer initialization in case of subsampling. 

stages (santander dataset) | before https://github.com/dmlc/xgboost/pull/6648 | thread safe https://github.com/dmlc/xgboost/pull/6648 and master | current PR
--| -- | -- | --
full training |136s | 165s | 140s
PredictRaw | 55s | 78s | 59s

this PR introduces threading for local buffers initialization.
But as still some gaps are exist is it possible to have not thread safe `PredictDMatrix` call during training?