@trivialfis 2021-02-09
I thought we have prediction cache enabled?