@ShvetsKS 2021-02-09
> I thought we have prediction cache enabled?

For subsampling case we still don't have prediction caching (seems it's still in progress: https://github.com/dmlc/xgboost/pull/6683, major complexity is related to multiclass classification case as for each group own indices subset is generated.)