@trivialfis 2021-02-10
@ShvetsKS Would you like to take a look into the GPU implementation of subsampling?  Or the WIP rewrite for CPU hist?  The subsampling doesn't have to conflict with cache.