@ShvetsKS 2021-02-10
> Last time your found it to be slower than master branch, which is expected as I don't have time to incooperate many recent optimization into it. But I believe my rewrite can at least offer some insight on how to implement some of the features in a different way.

Current changes are applicable even for inference stage as it's better to initialize `RegTree::FVec` thread locally anyway.