@ShvetsKS 2021-02-10
problem with continuous-integration/travis-ci/pr  is related with server node
`await ServerNode.close(self)`
seems restarting should help.